Additional help text and accompanying test; Relocate help test to CLI unit.
